+++ modules/common/automation.nix
@@ -14,8 +14,8 @@ in
         enable = true;
         flake = "git+ssh://git@github.com/caelx/nixos-config.git?ref=main";
         flags = [
-          "--update-input" "nixpkgs"
-          "--commit-lock-file"
+          "--update-input"
+          "nixpkgs"
         ];
         dates = "04:00";
         randomizedDelaySec = "45min";